Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/open-script-src-02-expected.txt

CONSOLE MESSAGE: line 5: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-inspan.pl?q=%3Cscript%20src=http://127.0.0.1:8000/security/xssAuditor/resources/xss.js?' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
